add INFO packet

+++ packets/server/collections_points.go
@@ -4,5 +4,5 @@ const COLLECTIONS_POINTS PacketServerType = 130
 
 // COLLECTIONS_POINTS(130) "I"
 type CollectionsPoints struct {
-	Points int
+	Points uint16
 }
